+++ packages/react-core/src/index.d.ts
@@ -19,6 +19,8 @@ export { groupValuesByColumn } from './operations/groupby';
 export { histogram } from './operations/histogram';
 export { scatterPlot } from './operations/scatterPlot';
 
+export { FilterTypes as _FilterTypes } from './filters/FilterQueryBuilder';
+
 export { tileFeatures } from './filters/tileFeatures';
 export { geojsonFeatures } from './filters/geojsonFeatures';